Separate file for DBIterator and related types
